@@ -2,27 +2,32 @@
   "name": "jssimplifyhtmltable",
   "version": "0.0.0",
   "description": "",
+  "type": "module",
   "main": "dist_nodejs/index.js",
   "types": "dist_nodejs/index.d.ts",
   "files": [
     "/dist_nodejs"
   ],
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
     "start": "webpack serve --open --config webpack.dev.js",
     "tsc": "tsc",
     "build": "webpack --config webpack.prod.js && npm run tsc"
   },
   "dependencies": {
-    "jssimplifytable": "git+https://github.com/ValorNaram/js-simplifyTable.git"
+    "jssimplifytable": "github:ValorNaram/js-simplifyTable#v1.0.0"
   },
   "devDependencies": {
+    "@jest/globals": "^30.0.4",
     "@types/node": "^20.7.1",
     "copy-webpack-plugin": "^12.0.2",
     "html-webpack-plugin": "^5.5.3",
+    "jest": "^30.0.5",
+    "jest-environment-jsdom": "^30.0.4",
+    "ts-jest": "^29.4.0",
     "ts-loader": "^9.4.2",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.5",
+    "typescript": "^5.3",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.1.4",
     "webpack-dev-middleware": "*",
